🚚[Rename] bookclub 타입 파일명 수정 #175
