+++ src/features/bookclub/api/bookclubApi.ts
@@ -1,5 +1,5 @@
 import apiClient from '@/lib/utils/apiClient';
-import { BookClubParams } from '../types';
+import { BookClubParams } from '../types/bookclubs';
 
 export const getBookClubs = async (params?: BookClubParams) => {
   const response = await apiClient.get('/book-clubs', { params });
